refactor: move dark theme helper to shared module
